[neptune/#112] Field Visit workflow - Assess & Post-Maintenance Assess Bring over controller actions for the assessment observation types and do some of the strategic Find+Replaces. Doesn't work yet.
